+++ src/components/genre-question-screen/genre-question-screen.js
@@ -2,7 +2,7 @@ import React, {useState} from 'react';
 import PropTypes from 'prop-types';
 import {Redirect} from 'react-router-dom';
 import {connect} from 'react-redux';
-import {incStepAction, incMistakesAction} from '../../action-creators/action-creators';
+import ActionCreator from '../../store/action-creator/action-creator';
 import {checkAnswers} from '../../helpers';
 import AudioPlayer from '../audio-player/audio-player';
 import GameMistakes from '../game-mistakes/game-mistakes';
@@ -106,9 +106,9 @@ const mapStateToProps = (state) => ({
 const mapDispatchToProps = (dispatch) => ({
   onAnswerClick: (type, answer) => {
     if (!checkAnswers(type, answer)) {
-      dispatch(incMistakesAction());
+      dispatch(ActionCreator.increaseMistakesAction());
     } else {
-      dispatch(incStepAction());
+      dispatch(ActionCreator.increaseStepAction());
     }
   }
 });
