+++ src/index.js
@@ -4,10 +4,10 @@ import {createStore, applyMiddleware} from 'redux';
 import {Provider} from 'react-redux';
 import thunk from 'redux-thunk';
 import {composeWithDevTools} from 'redux-devtools-extension';
-import {reducer} from './reducers/reducer';
+import {reducer} from './store/reducers/reducer';
 import App from './components/app/app';
 import {createAPI} from './api';
-import {Operation as DataOperation} from './reducers/questionsReducer/questionsReducer';
+import DataOperation from './store/operations/data-operation';
 // import questions from './mocks/questions';
 // import settings from './mocks/settings';
