+++ src/store/action-types/action-types.js
@@ -1,9 +1,11 @@
 const INC_MISTAKES = `INC_MISTAKES`;
 const INC_STEP = `INC_STEP`;
 const RESTART_GAME = `RESTART_GAME`;
+const LOAD_QUESTIONS = `LOAD_QUESTIONS`;
 
 export {
   INC_MISTAKES,
   INC_STEP,
-  RESTART_GAME
+  RESTART_GAME,
+  LOAD_QUESTIONS
 };
