+++ src/store/reducers/mistakesReducer/mistakesReducer.js
@@ -1,4 +1,4 @@
-import {INC_MISTAKES, RESTART_GAME} from '../../actions/actions';
+import {INC_MISTAKES, RESTART_GAME} from '../../action-types/action-types';
 
 function mistakesReducer(state = 0, action) {
   switch (action.type) {
