+++ src/store/reducers/stepReducer/stepReducer.test.js
